backend/requirements.txt
flask
flask-cors  
mp-api
python-dotenv
